### 2017-09-05 #390 @whedon
[PRE REVIEW]: ResumableFunctions: C# sharp style generators for Julia.
**Submitting author:** @benlauwens (<a href="http://orcid.org/0000-0003-0761-6265">Ben Lauwens</a>)
**Repository:** <a href="https://github.com/BenLauwens/ResumableFunctions.jl.git" target ="_blank">https://github.com/BenLauwens/ResumableFunctions.jl.git</a>
**Version:** v0.1.0
