### 2017-09-05 #390 @danielskatz
I have no experience with c# or julia, but I can try to edit this.

@BenLauwens - please let me know if you have reviewer suggestions.
