### 2017-09-05 aee6770
Fix bug in get_slots
