+++ src/utils.jl
@@ -6,10 +6,11 @@ Function returning the slots of a function definition
 """
 function get_slots(func_def::Dict) :: Dict{Symbol, Type}
   slots = Dict{Symbol, Type}()
-  func_name = gensym()
-  func_def[:name] = func_name
-  eval(combinedef(func_def |> flatten))
-  code_data_infos = @eval code_typed($func_name)
+  func_def[:name] = gensym()
+  func_expr = combinedef(func_def) |> flatten
+  func = Main.eval(func_expr)
+  println(func)
+  code_data_infos = @eval code_typed(Main.$(func_def[:name]))
   (code_info, data_type) = code_data_infos[1]
   for (i, slotname) in enumerate(code_info.slotnames)
     slots[slotname] = code_info.slottypes[i]
