### 2017-09-06 #390 @BenLauwens
Hi @danielskatz

I propose three reviewers:
- Dr [Davide Lasagna ](davide.lasagna@soton.ac.uk): he knows Julia and he is interested to use `ResumableFunctions` to express complex iteration patterns
- Assoc Prof [Bart Janssens](Bart.janssens@rma.ac.be): he is a colleague who has developed `QML.jl` and needs the functionalities of `ResumableFunctions` to reduce the overhead of progress bar updates in a GUI
- [Jeffrey Sarnoff](jeffrey.sarnoff@gmail.com): during JuliaCon2016 he was hosting the session in which I presented an extension of `SimJulia`. He has developed several Julia package and knows the language and its tools very well

They are all familiar with Julia and the first two have a clear use case for `ResumableFunctions` that is complementary to my use case (process description in an event driven simulator).

Kind regards

Ben
