### 2017-09-06 8fab4fa
Fix bug in transform slots
