### 2017-09-06 #390 @danielskatz
@gasagna @barche @Jeffrey-Sarnoff - would one of you be willing to review this software paper for [JOSS](http://joss.theoj.org)?

We have some [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) to let you know what we expect of a reviewer/review.

Reviews are done through github issues - once a reviewer is assigned, I will close this issue and open a review issue, that you will go through as you review the paper and the software.  An example is https://github.com/openjournals/joss-reviews/issues/357 (this paper was just accepted.)
