+++ src/macro.jl
@@ -21,9 +21,11 @@ Macro that transforms a function definition in a finite-statemachine:
 
 macro resumable(expr::Expr)
   expr.head != :function && error("Expression is not a function definition!")
+  args = get_arguments(expr)
+  shift!(args)
   func_def = splitdef(expr)
-  args = [(begin (a, b, c, d) = splitarg(arg); :($a) end for arg in func_def[:args])...,
-          (begin (a, b, c, d) = splitarg(arg); :($a) end for arg in func_def[:kwargs])...]
+  #args = [(begin (a, b, c, d) = splitarg(arg); :($a) end for arg in func_def[:args])...,
+  #        (begin (a, b, c, d) = splitarg(arg); :($a) end for arg in func_def[:kwargs])...]
   ui8 = BoxedUInt8(zero(UInt8))
   func_def[:body] = postwalk(x->transform_for(x, ui8), func_def[:body])
   slots = get_slots(copy(func_def))
