### 2017-09-07 facc4ef
Optimize get_arg_name
