### 2017-09-08 944a554
Clean up macrotoolutils
