+++ src/macrotoolutils.jl
@@ -3,32 +3,72 @@
 
 using MacroTools: @q, striplines, @match, @capture
 
+isshortdef(ex) = (@capture(ex, (fcall_ = body_)) &&
+                  (@capture(gatherwheres(fcall)[1],
+                            (f_(args__) |
+                             f_(args__)::rtype_))))
+
 function longdef1(ex)
+  if @capture(ex, (arg_ -> body_))
+    @q function ($arg,) $body end
+  elseif isshortdef(ex)
+    @assert @capture(ex, (fcall_ = body_))
+    striplines(Expr(:function, fcall, body))
+  else
+    ex
+  end
+end
+
+function shortdef1(ex)
   @match ex begin
-    (f_(args__) = body_) => @q function $f($(args...)) $body end
-    (f_(args__)::rtype_ = body_) => @q function $f($(args...))::$rtype $body end
-    ((args__,) -> body_) => @q function ($(args...),) $body end
-    (arg_ -> body_) => @q function ($arg,) $body end
-    (f_(args__) where {whereparams__} = body_) =>
-        @q function $f($(args...)) where {$(whereparams...)}
-            $body end
-    ((f_(args__)::rtype_) where {whereparams__} = body_) =>
-        @q function ($f($(args...))::$rtype) where {$(whereparams...)}
-            $body end
+    function f_(args__) body_ end => @q $f($(args...)) = $body
+    function f_(args__)::rtype_ body_ end => @q $f($(args...))::$rtype = $body
+    function (args__,) body_ end => @q ($(args...),) -> $body
+    ((args__,) -> body_) => ex
+    (arg_ -> body_) => @q ($arg,) -> $body
     _ => ex
   end
 end
 
+""" `gatherwheres(:(f(x::T, y::U) where T where U)) => (:(f(x::T, y::U)), (:U, :T))`
+"""
+function gatherwheres(ex)
+  if @capture(ex, (f_ where {params1__}))
+    f2, params2 = gatherwheres(f)
+    (f2, (params1..., params2...))
+  else
+    (ex, ())
+  end
+end
+
+doc"""    splitdef(fdef)
+Match any function definition
+```julia
+function name{params}(args; kwargs)::rtype where {whereparams}
+   body
+end
+```
+and return `Dict(:name=>..., :args=>..., etc.)`. The definition can be rebuilt by
+calling `MacroTools.combinedef(dict)`, or explicitly with
+```
+rtype = get(dict, :rtype, :Any)
+all_params = [get(dict, :params, [])..., get(dict, :whereparams, [])...]
+:(function $(dict[:name]){$(all_params...)}($(dict[:args]...);
+                                            $(dict[:kwargs]...))::$rtype
+      $(dict[:body])
+  end)
+```
+"""
 function splitdef(fdef)
   error_msg = "Not a function definition: $fdef"
   @assert(@capture(longdef1(fdef),
-                   function ((fcall_ where {whereparams__}) | fcall_)
-                   body_ end),
-          error_msg)
-  @assert(@capture(fcall, ((func_(args__; kwargs__)) |
-                           (func_(args__; kwargs__)::rtype_) |
-                           (func_(args__)) |
-                           (func_(args__)::rtype_))),
+                   function (fcall_ | fcall_) body_ end),
+          "Not a function definition: $fdef")
+  fcall_nowhere, whereparams = gatherwheres(fcall)
+  @assert(@capture(fcall_nowhere, ((func_(args__; kwargs__)) |
+                                   (func_(args__; kwargs__)::rtype_) |
+                                   (func_(args__)) |
+                                   (func_(args__)::rtype_))),
           error_msg)
[…]
