### 2017-09-14 89ae538
Update manual.md
