### 2017-09-14 8f2ec19
Fix bug in try body__ end transformation
