+++ v0.1.4/search_index.js
@@ -1,0 +1,419 @@
+var documenterSearchIndex = {"docs": [
+
+{
+    "location": "index.html#",
+    "page": "Home",
+    "title": "Home",
+    "category": "page",
+    "text": ""
+},
+
+{
+    "location": "index.html#ResumableFunctions-1",
+    "page": "Home",
+    "title": "ResumableFunctions",
+    "category": "section",
+    "text": "C# sharp style generators a.k.a. semi-coroutines for Julia.C# has a convenient way to create iterators using the yield return statement. The package ResumableFunctions provides the same functionality for the Julia language by introducing the @resumable and the @yield macros. These macros can be used to replace the Task switching functions produce and consume which were deprecated in Julia v0.6. Channels are the preferred way for inter-task communication in julia v0.6+, but their performance is subpar for iterator applications."
+},
+
+{
+    "location": "index.html#Example-1",
+    "page": "Home",
+    "title": "Example",
+    "category": "section",
+    "text": "using ResumableFunctions\n\n@resumable function fibonnaci(n::Int)\n  a = 0\n  b = 1\n  for i in 1:n-1\n    @yield a\n    a, b = b, a+b\n  end\n  a\nend\n\nfor val in fibonnaci(10) \n  println(val) \nend\n\n# output\n\n0\n1\n1\n2\n3\n5\n8\n13\n21\n34"
+},
+
+{
+    "location": "index.html#Installation-1",
+    "page": "Home",
+    "title": "Installation",
+    "category": "section",
+    "text": "ResumableFunctions is a registered package and can be installed by running:Pkg.add(\"ResumableFunctions\")"
+},
+
+{
+    "location": "index.html#Authors-1",
+    "page": "Home",
+    "title": "Authors",
+    "category": "section",
+    "text": "Ben Lauwens, Royal Military Academy, Brussels, Belgium."
+},
+
+{
+    "location": "index.html#License-1",
+    "page": "Home",
+    "title": "License",
+    "category": "section",
+    "text": "ResumableFunctions is licensed under the MIT \"Expat\" License."
+},
+
+{
+    "location": "manual.html#",
+    "page": "Manual",
+    "title": "Manual",
+    "category": "page",
+    "text": ""
+},
+
+{
+    "location": "manual.html#Manual-1",
+    "page": "Manual",
+    "title": "Manual",
+    "category": "section",
+    "text": ""
+},
+
+{
+    "location": "manual.html#Basic-usage-1",
+    "page": "Manual",
+    "title": "Basic usage",
+    "category": "section",
[…]
