### 2017-09-23 #400 @danielskatz
The only time I've heard of this problem, it was when someone tried to check a box when they were not logged in.

So, please be sure you are logged in when you try to check a box.  (I know you were when you commented, but perhaps you were not initially?)

Also, when you say "edit the checklist", do you really mean check a box?  You should be able to do that, but you should not able to make other changes to the checklist.

Let me know if either of these clear up the problem.
