### 2017-09-29 #400 @gasagna
No, sorry, I can't really check the boxes. I am logged in. Tried two different browsers as well.
