### 2017-09-29 #400 @danielskatz
@arfon - any thoughts on this?  I haven't seen this before, and you are probably the GitHub expert :)
