### 2017-10-01 #400 @arfon
> No, sorry, I can't really check the boxes. I am logged in. Tried two different browsers as well.

Looks like the API command to give you access must have failed silently. @gasagna - can you visit this URL (logged in): https://github.com/openjournals and accept the invite. You should then be able to update the check list.
