### 2017-10-02 #400 @arfon
> Sorry, still unable to check the boxes after the invite.

OK, I officially give up. Please copy and paste the checklist content into a separate comment in this thread and work through it here.
