### 2017-10-09 BenLauwens/ResumableFunctions.jl#3 @gasagna
`@resumable` functions do not support `where` type specifications
This code

```julia
using ResumableFunctions

f(x) = 0.5*x*(x-1)

@resumable function foo(x::T) where {T}
    for i = 1:10
        @yield x
        x = f(x)
    end
    x
end
```

results in

```
ERROR: LoadError: UndefVarError: T not defined
Stacktrace:
 [1] include_from_node1(::String) at ./loading.jl:576
 [2] include(::String) at ./sysimg.jl:14
 [3] process_options(::Base.JLOptions) at ./client.jl:305
 [4] _start() at ./client.jl:371
while loading ../tests.jl, in expression starting on line 32
```

This issue is not blocking openjournals/joss-reviews#400 to be accepted, but at least it should be mentioned in the caveats. In the long term, supporting `where` type specifications would be a very useful feature to have.
