### 2017-10-09 BenLauwens/ResumableFunctions.jl#4 @gasagna
iterator interface semantics
I am testing the functionality of this repo and I stumbled in a silly usability issue. Not really an issue, but the behaviour could be documented better. Consider this code:

```julia
using ResumableFunctions

f(x) = 0.5*x*(x-1)

@resumable function foo(x)
    out = f(x)
    for i = 1:4
        @yield out
        out = f(out)
    end
    out
end

t = foo(0.6)

for el in t
    println(el)
end

for el in t
    println(el)
end
```

Only five numbers are printed in this code, because the second loop does not print any. Naively, I would have expected to have the same sequence  printed twice, but this does not happen because the iterator state is hidden inside the resumable function definition and does not get reinitialised when the relevant `start` method is called.

This makes the semantics of the iterator interface slightly different than other iterators in Julia (e.g. zip([1, 2, 3], [4, 5, 6])). This is perfectly OK for openjournals/joss-reviews#400 to proceed, but it should be documented, e.g. in the Iterator Interface section of the docs.
