### 2017-10-09 BenLauwens/ResumableFunctions.jl#3 @gasagna
I now see in the readme that parametric resumable functions are not currently supported. Apologies. This could go in the docs as well. How hard is to implement this functionality?
