### 2017-10-09 #400 @gasagna
Hi Daniel,

I have completed my review, and I have pointed out minor issues that @BenLauwens should address (see unchecked boxes above). These are mostly documentation improvement suggestions.

I will wait for the author to fix these and then check the remaining boxes.
