### 2017-10-15 BenLauwens/ResumableFunctions.jl#4 @BenLauwens
I see your point. I will update the docs.
