### 2017-10-15 34b407a
Fix issue 4
