### 2017-10-15 ab2adb3
Document a caveat when reusing a FiniteStateMachineIterator object in multiple for loops
