### 2017-10-15 BenLauwens/ResumableFunctions.jl#4 @BenLauwens
I have modified the `start` method so that the semantics are the same as the other iterators in Julia. There is a caveat that only the `state` variable is reinitialised but not the other internal variables. So a `@resumable function` that alters the values of its arguments will show some unexpected behaviour:
```Julia
julia> @resumable function fib(a, b)
       for i in 1:4
           @yield a
           a, b = b, a+b
       end
       a
       end

fib (generic function with 1 method)

julia> fib_it = fib(0,1)
##660(0x00, 0, 1, 4603404944, 4603404464:257, 8)

julia> for el in fib_it
       println(el)
       end
0
1
1
2
3

julia> for el in fib_it
       println(el)
       end
3
5
8
13
21
```

I added a comment to the caveats section.
