### 2017-10-15 1952387
Add todo for parametric resumable functions in docs
