### 2017-10-15 BenLauwens/ResumableFunctions.jl#3 @BenLauwens
I am waiting for a new release of `Macrotools` to implement this functionality. I have only to change the type definition and the constructor but this is fairly trivial once `where` is completely supported in `Macrotools`. I have added a todo section to the docs.
