### 2017-10-24 #400 @BenLauwens
Hi @danielskatz. I updated the documentation as proposed by @gasagna and I modified one line of code so that the behaviour of the iterator created by an `@resumable function` is in line with the already existing iterators in Julia.
One issue remains that is documented both in the README and the tutorial. Parametrised `@resumable functions` are not supported yet. The implementation of this feature is straightforward but tedious. I am using a supporting package Macrotools to do the heavy work but this package has not yet released a version that supports the 'where' clause (in master this is already supported). This is however a minor feature and ResumableFunctions are perfectly working without it.
