### 2017-10-27 0ca36ae
Remove macrotoolutils, functionality is no included in MacroTools
