### 2017-10-27 BenLauwens/ResumableFunctions.jl#4 @BenLauwens
I just released a new version of ResumableFunctions including this modification. So I will close this issue soon. Comments are always welcome.
