### 2017-10-29 #400 @gasagna
I think this is OK. There are a few missing checks in my reviews, but these are not blocking acceptance. Just additional edits to the docs that the author might wish to add.
