### 2017-10-30 #400 @danielskatz
@gasagna - we do want to see all the boxes checked before we accept.  For example, for performance, the description is
> If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

So please check off everything you can, and if you don't feel you can check something off, please explain what you think is needed in order that you could do so.
