### 2017-10-30 #400 @gasagna
@danielskatz The text in italic in my check list explains it already. My comments are suggestions to improve the docs that the author might wish to take into account. These are not acceptance-blocking.
