### 2017-10-30 #400 @danielskatz
Thanks @gasagna, but as I said, we will not move ahead until all boxes are checked.  Can I suggest that rather than altering the list, you check the items (since you say that they are not blockers) and make the comments in italics issues that you open in the author's package?

Remember that if this was a normal review (where you could check the boxes on the top of the issue), you would not be able to edit the checklist, just check the boxes or leave the empty.
