### 2017-10-30 be955a6
Removing # from bibtex string
