+++ paper/paper.bib
@@ -14,7 +14,7 @@
   urldate = {2017-09-01}
 }
 
-#article{Julia,
+@article{Julia,
   author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
   title = {Julia: A Fresh Approach to Numerical Computing.},
   year = 2017,
@@ -25,10 +25,10 @@
   doi={https://doi.org/10.1137/141000671}
 }
 
-#online{C#Iterators,
-  author = {Microsoft}
+@online{CsharpIterators,
+  author = {Microsoft},
   title = {Iterators (C#).},
-  year = 2017
+  year = 2017,
   url = {https://docs.microsoft.com/en-us/dotnet/csharp/programming-guide/concepts/iterators},
-  urldate = {2017-09-01} 
+  urldate = {2017-09-01}
 }
