### 2017-10-30 #400 @arfon
@BenLauwens - please merge this PR with some minor changes to your paper and bibtex https://github.com/BenLauwens/ResumableFunctions.jl/pull/8

Then, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
