+++ versions.js
@@ -2,6 +2,7 @@ var DOC_VERSIONS = [
   "stable",
   "latest",
   "release-0.1",
+  "v0.1.6",
   "v0.1.5",
   "v0.1.4",
   "v0.1.3",
