### 2017-10-30 #400 @BenLauwens
@arfon
I have just merged the changes and created a new release.
DOI: 10.5281/zenodo.1039345
